refactor: Move `.../test_helpers.rs` to `.../test/mod.rs`
